Kotlin KSP Nullability Fix
When using nullable types in JPA Entity fields, the type determination code was throwing exceptions. This change creates a local version of the KSType.toClassName() function which gets the ClassName while ignoring the KSType's nullability and parameterization. It also ensures that the ClassName returned does not include a nullability modifier as it previously did with Enums.